requirements.txt
pinecone-client
sentence-transformers
llama-cpp-python
